app/pyproject.toml
[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.poetry]
name = "EthBotDetector"
version = "0.0.1"
description = ""
authors = ["Jan Romaniak"]

[tool.poetry.dependencies]
python = "3.9.4"
faust = "^1.10.4"
simple-settings = "^1.0.0"
watchdog = "^2.0.3"
PyYAML = "^5.4.1"
argh = "^0.26.2"
requests = "^2.25.1"
web3 = "^5.19.0"
pyfcm = "^1.5.1"
psycopg2 = "^2.8.6"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.2"}
asyncpg = "^0.22.0"
